# docs/openapi/event.yaml
openapi: 3.1.0
info:
  title: Social Event Mapper Event API
  version: 0.1.0
  description: |
    Authenticated event-management endpoints for creating and managing social events.

    Obtain an access token from the Auth API (`POST /auth/login` or
    `POST /auth/register/email/verify`) and send it as `Authorization: Bearer <token>`
    when calling protected operations in this definition.

    Behavioral notes for create-event clients:
    - The authenticated user becomes the event host.
    - A host can create at most one event with a given `title`. Reusing the same trimmed title for the same host returns `409 Conflict`.
    - Tags and constraints are persisted in the same transaction as the event.
    - Enum-like fields are case-sensitive and must be sent in uppercase.
    - `start_time` and `end_time` must be RFC 3339 date-times with a timezone designator such as `Z` or `+03:00`.
    - Each event has **exactly one** category: send a single `category_id` (not a list). It must reference an existing row in `event_category`. The default database seed defines twenty categories (ids `1`–`20`); see `CreateEventRequestBase.category_id` for the id-to-name map.
servers:
  - url: /api
...
paths:
...
  /me/favorites:
    get:
      tags: [Events]
      summary: List favorite events
      description: Returns the authenticated user's favorite events ordered by most recently favorited first.
      operationId: listFavoriteEvents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Favorite events listed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteEventsResponse'
        '401':
          description: Missing or invalid access token.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    FavoriteEventsResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FavoriteEventItem'
    FavoriteEventItem:
      type: object
      additionalProperties: false
      required: [id, title, status, privacy_level, location_address, start_time, favorited_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        category:
          type:
            - string
            - "null"
        image_url:
          type:
            - string
            - "null"
        status:
          type: string
          description: Event status, for example `ACTIVE`, `IN_PROGRESS`, `CANCELED`, or `COMPLETED`.
          example: ACTIVE
        privacy_level:
          type: string
          enum: [PUBLIC, PROTECTED, PRIVATE]
          description: Current stored event visibility level.
          example: PUBLIC
        location_address:
          type:
            - string
            - "null"
          description: Optional human-readable location address stored for the event.
          example: Kadikoy, Istanbul
        start_time:
          type: string
          format: date-time
        end_time:
          type:
            - string
            - "null"
          format: date-time
        favorited_at:
          type: string
          format: date-time